Fixed model training/evaluation mode bug
PyTorch models are created in "traing" mode by default, but the model
should be switched to "evaluation" mode at the appropriate
times. Depending on the mode of the model, PyTorch does a lot of fancy
math behind the scenes. The difference between "training" and
"evaluation" mode is the same for the `Linear` model because this class
is relatively simple. I've added a few lines to make the `Runner` class
more general, so it can hanndle `torch.nn.Dropout`, `torch.nn.BatchNorm`
and other PyTorch modules that are sensitive to the model's mode.